% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dark}
\alias{dark}
\title{Object defining the app's dark mode ui layout}
\format{
\subsection{\verb{R object}}{

An object describing the ui dark mode's colour patterns
\describe{
\item {dark}{a bstheme object containing information on theme and colours of the application's dark mode}
}
}
}
\usage{
toggle switch in the ui
}
\description{
Object defining the app's dark mode ui layout
}
\keyword{datasets}
