% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_information_integration.R
\name{gene_information_integration}
\alias{gene_information_integration}
\title{Gene visualisation functions}
\usage{
gene_information_integration(res)
}
\arguments{
\item{res}{a specific gene containing variants}
}
\value{
an image displaying the positioning of the variants relative to the gene transcripts
}
\description{
Gene visualisation functions
}
