% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{get_light_theme}
\alias{get_light_theme}
\alias{get_dark_theme}
\title{Application Themes}
\usage{
get_light_theme()

get_dark_theme()
}
\description{
Define the light and dark themes for the VariantsInvestigator application.
These functions return \code{bslib::bs_theme} objects.
}
