% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{light}
\alias{light}
\title{Object defining the app's light mode ui layout}
\format{
\subsection{\verb{R object}}{

An object describing the ui light mode's colour patterns
\describe{
\item{light}{a bstheme object containing information on theme and colours of the application's light mode}
}
}
}
\usage{
toggle switch in the ui
}
\description{
Object defining the app's light mode ui layout
}
\keyword{datasets}
