% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refgenome_model}
\alias{refgenome_model}
\title{Dataset containing exome references for gene matching}
\format{
\subsection{\code{tibble}}{

A tibble with 10 columns
\describe{
\item{chromosome}{the chromosome containing the feature}
\item{start}{start position of the feature}
\item{end}{end position of the feature}
\item{width}{number of nucleotides of the feature}
\item{strand}{strand on which the feature is positioned}
\item{feature}{type of data regarding the gene, either transcript, exon or gene}
\item{gene}{Ensembl's standard gene reference}
\item{exon}{Ensembl's standard exon reference}
\item{transcript}{Ensembl's standard transcript reference}
\item{symbol}{gene name}
}
}
}
\source{
created with TxDb.Hsapiens.UCSC.hg19.knownGene R package
}
\usage{
refgenome_model
}
\description{
Dataset containing exome references for gene matching
}
\keyword{datasets}
