% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_dataset}
\alias{sample_dataset}
\title{Sample Dataset}
\format{
\subsection{\code{sample_dataset}}{

A tibble with 33 columns:
\describe{
\item{variable}{The name of the predictor}
\item{description}{The description of the predictor}
}

@source generated with simstudy
}

\subsection{\code{sample_dataset}}{

A tibble with 33 columns:
\describe{
\item{variable}{The name of the predictor}
\item{description}{The description of the predictor}
}

@source generated with simstudy
}

\subsection{\code{sample_dataset}}{

A tibble with 33 columns:
\describe{
\item{variable}{The name of the predictor}
\item{description}{The description of the predictor}
}

@source generated with simstudy
}
}
\usage{
data(sample_dataset)

data(sample_dataset)

data(sample_dataset)
}
\description{
Dataset used to test the functions in the package
containing simulated data

Dataset used to test the functions in the package
containing simulated data

Dataset used to test the functions in the package
containing simulated data
}
\keyword{datasets}
