% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantsInvestigator.R
\name{VariantsInvestigator}
\alias{VariantsInvestigator}
\title{VariantsInvestigator app core}
\usage{
VariantsInvestigator()
}
\value{
A Shiny app object
}
\description{
Launches the main Shiny application for investigating variants from the generated SQLite database.
}
