% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfToSQL.R
\name{vcfToSQL}
\alias{vcfToSQL}
\title{VCFtoSQL shiny app}
\usage{
vcfToSQL(...)
}
\arguments{
\item{.vcf.gz}{}
}
\value{
an SQL database containing the VCF informations split in two tables containing the data, namely
genotype and variants, and a third table containing rejected lines with the reason of the rejection
}
\description{
VCFtoSQL shiny app
}
