% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantsInvestigator.R
\name{VariantsInvestigator}
\alias{VariantsInvestigator}
\title{VariantsInvestigator app core}
\usage{
VariantsInvestigator(...)
}
\arguments{
\item{.sqlite}{}
}
\value{
a dataset of filtered variants
}
\description{
VariantsInvestigator app core
}
