% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfToSQL.R
\name{vcfToSQL}
\alias{vcfToSQL}
\title{VCFtoSQL shiny app}
\usage{
vcfToSQL()
}
\value{
A Shiny app object
}
\description{
Launches a Shiny application to convert a VCF file into an SQLite database.
The database will contain tables for variants, genotypes, and rejected lines.
}
